fix page and organisms
